requirement.txt
aliyun-python-sdk-core==2.13.13
aliyun-python-sdk-dysmsapi==1.0.0
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.0.2
django-debug-toolbar==2.2
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
idna==2.8
jieba==0.39
jmespath==0.9.4
mysqlclient==1.4.6
Pillow==7.0.0
python-memcached==1.59
pytz==2019.3
qiniu==7.2.0
requests==2.22.0
shortuuid==0.5.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
Whoosh==2.7.4
